071 docker-compose for local dev

+++ src/main/java/ru/questsfera/questreservation/cache/repository/ReservationCacheRepository.java
@@ -9,7 +9,7 @@ import java.time.LocalTime;
 import java.util.Optional;
 
 @Repository
-public interface ReservationCacheRepository extends CrudRepository<ReservationCache, Integer> {
+public interface ReservationCacheRepository extends CrudRepository<ReservationCache, Long> {
 
     Optional<ReservationCache> findByQuestIdAndDateReserveAndTimeReserve(
             Integer questId,
